# docs/swagger.yaml
openapi: "3.0.3"
info:
  title: "Course & Training Management System API (ИПР №1)"
  version: "1.0.0"
  description: "Безопасный REST API для системы управления курсами, личным кабинетом студента, кабинетом преподавателя и динамическим RBAC (Role-Based Access Control). Спецификация сгенерирована из декларативных аннотаций Swaggo (// Method Godoc, @Summary, @Description, @Tags, @Param, @Success, @Failure, @Router) в контроллерах."
servers:
  - url: "/api/v1"
    description: "API v1 Base Endpoint"
...
components:
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
      description: "Введите JWT Access токен (Bearer <token>)"
  schemas:
    RegisterRequest:
      type: "object"
      required:
        - "email"
        - "password"
        - "full_name"
        - "role"
      properties:
        email:
          type: "string"
          format: "email"
          example: "student@example.com"
        password:
          type: "string"
          minLength: 8
          example: "StudentSecurePass123!"
        full_name:
          type: "string"
          example: "Иван Иванов"
        role:
          type: "string"
          enum:
            - "student"
            - "instructor"
            - "manager"
          example: "student"
          description: "Обязательная роль пользователя (student, instructor, manager)"
      example:
        email: "student@example.com"
        password: "StudentSecurePass123!"
        full_name: "Иван Иванов"
        role: "student"
    LoginRequest:
      type: "object"
      required:
        - "email"
        - "password"
      properties:
        email:
          type: "string"
          format: "email"
          example: "admin@course-platform.local"
        password:
          type: "string"
          example: "AdminSecurePass123!"
      example:
        email: "admin@course-platform.local"
        password: "AdminSecurePass123!"
...
    TokenResponse:
      type: "object"
      properties:
        success:
          type: "boolean"
          example: true
        data:
          type: "object"
          properties:
            access_token:
              type: "string"
              example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      example:
        success: true
        data:
          access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiJ1c3ItMSIsInJvbGUiOiJzdHVkZW50In0..."
    AuthResponse:
      type: "object"
      properties:
        success:
          type: "boolean"
          example: true
        message:
          type: "string"
          example: "Аутентификация успешна."
        data:
          type: "object"
          properties:
            access_token:
              type: "string"
              example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            user:
              type: "object"
              required:
                - "id"
                - "email"
                - "full_name"
                - "role"
              properties:
                id:
                  type: "string"
                  example: "usr-uuid-1"
                email:
                  type: "string"
                  example: "admin@course-platform.local"
                full_name:
                  type: "string"
                  example: "Администратор Системы"
                role:
                  type: "object"
                  required:
                    - "id"
                    - "name"
                    - "permissions"
                  properties:
                    id:
                      type: "string"
                      example: "role-uuid-admin"
                    name:
                      type: "string"
                      example: "admin"
                    description:
                      type: "string"
                      example: "Полный администратор платформы"
                    permissions:
                      type: "array"
                      items:
                        type: "string"
                      example:
                        - "users:view_all"
                        - "users:manage_roles"
                        - "roles:manage"
                        - "courses:create"
                        - "payments:view_all"
      example:
        success: true
        message: "Аутентификация успешна."
        data:
          access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiJ1c3ItMSIsInJvbGUiOiJhZG1pbiJ9..."
          user:
            id: "usr-uuid-1"
            email: "admin@course-platform.local"
            full_name: "Администратор Системы"
            role:
              id: "role-uuid-admin"
              name: "admin"
              description: "Полный администратор платформы"
              permissions:
                - "users:view_all"
                - "users:manage_roles"
                - "roles:manage"
                - "courses:create"
                - "payments:view_all"
    CurrentUserResponse:
      type: "object"
      properties:
        success:
          type: "boolean"
          example: true
        data:
          type: "object"
          properties:
            id:
              type: "string"
              example: "usr-uuid-1"
            email:
              type: "string"
              example: "admin@course-platform.local"
            full_name:
              type: "string"
              example: "Администратор Системы"
            role:
              type: "object"
              required:
                - "id"
                - "name"
                - "permissions"
              properties:
                id:
                  type: "string"
                  example: "role-uuid-admin"
                name:
                  type: "string"
                  example: "admin"
                description:
                  type: "string"
                  example: "Полный администратор платформы"
                permissions:
                  type: "array"
                  items:
                    type: "string"
                  example:
                    - "users:view_all"
                    - "roles:manage"
            created_at:
              type: "string"
              example: "2026-10-04T00:00:00.000Z"
      example:
        success: true
        data:
          id: "usr-uuid-1"
          email: "admin@course-platform.local"
          full_name: "Администратор Системы"
          role:
            id: "role-uuid-admin"
            name: "admin"
            description: "Полный администратор платформы"
            permissions:
              - "users:view_all"
              - "users:manage_roles"
              - "roles:manage"
              - "courses:create"
              - "payments:view_all"
          created_at: "2026-10-04T00:00:00.000Z"
...
    SuccessResponse:
      type: "object"
      properties:
        success:
          type: "boolean"
          example: true
        message:
          type: "string"
          example: "Операция успешно выполнена."
      example:
        success: true
        message: "Операция успешно выполнена."
    ErrorResponse:
      type: "object"
      properties:
        success:
          type: "boolean"
          example: false
        error:
          type: "object"
          properties:
            code:
              type: "string"
              example: "VALIDATION_ERROR"
            message:
              type: "string"
              example: "Неверные параметры запроса."
            details:
              type: "array"
              items:
                type: "object"
      example:
        success: false
        error:
          code: "VALIDATION_ERROR"
          message: "Неверные параметры запроса."
paths:
  /auth/register:
    post:
      tags:
        - "auth"
      summary: "Регистрация нового пользователя"
      description: "Регистрирует нового пользователя с обязательной ролью (student, instructor, manager) и хеширует пароль bcrypt"
      parameters: []
      responses:
        201:
          description: "models.AuthResponse"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
              example:
                success: true
                message: "Аутентификация успешна."
                data:
                  access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiJ1c3ItMSIsInJvbGUiOiJhZG1pbiJ9..."
                  user:
                    id: "usr-uuid-1"
                    email: "admin@course-platform.local"
                    full_name: "Администратор Системы"
                    role:
                      id: "role-uuid-admin"
                      name: "admin"
                      description: "Полный администратор платформы"
                      permissions:
                        - "users:view_all"
                        - "users:manage_roles"
                        - "roles:manage"
                        - "courses:create"
                        - "payments:view_all"
        400:
          description: "models.ErrorResponse"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                error:
                  code: "VALIDATION_ERROR"
                  message: "Неверные параметры запроса."
        409:
          description: "models.ErrorResponse"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                error:
                  code: "VALIDATION_ERROR"
                  message: "Неверные параметры запроса."
      requestBody:
        required: true
        description: "Данные регистрации с обязательной ролью"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            example:
              email: "student@example.com"
              password: "StudentSecurePass123!"
              full_name: "Иван Иванов"
              role: "student"
  /auth/login:
    post:
      tags:
        - "auth"
      summary: "Вход в систему"
      description: "Аутентификация с защитой от брутфорса (блокировка на 15 мин после 5 попыток)"
      parameters: []
      responses:
        200:
          description: "models.AuthResponse"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
              example:
                success: true
                message: "Аутентификация успешна."
                data:
                  access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiJ1c3ItMSIsInJvbGUiOiJhZG1pbiJ9..."
                  user:
                    id: "usr-uuid-1"
                    email: "admin@course-platform.local"
                    full_name: "Администратор Системы"
                    role:
                      id: "role-uuid-admin"
                      name: "admin"
                      description: "Полный администратор платформы"
                      permissions:
                        - "users:view_all"
                        - "users:manage_roles"
                        - "roles:manage"
                        - "courses:create"
                        - "payments:view_all"
        401:
          description: "models.ErrorResponse"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                error:
                  code: "VALIDATION_ERROR"
                  message: "Неверные параметры запроса."
        423:
          description: "models.ErrorResponse"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                error:
                  code: "VALIDATION_ERROR"
                  message: "Неверные параметры запроса."
      requestBody:
        required: true
        description: "Учетные данные"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              email: "admin@course-platform.local"
              password: "AdminSecurePass123!"
  /auth/refresh:
    post:
      tags:
        - "auth"
      summary: "Обновить Access Token"
      description: "Выпуск нового токена по Refresh Token из HttpOnly cookie"
      parameters: []
      responses:
        200:
          description: "models.TokenResponse"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
              example:
                success: true
                data:
                  access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiJ1c3ItMSIsInJvbGUiOiJzdHVkZW50In0..."
        401:
          description: "models.ErrorResponse"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                error:
                  code: "VALIDATION_ERROR"
                  message: "Неверные параметры запроса."
  /auth/logout:
    post:
      tags:
        - "auth"
      summary: "Выход из системы"
      description: "Очистка сессионных cookie"
      parameters: []
      responses:
        200:
          description: "models.SuccessResponse"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
              example:
                success: true
                message: "Операция успешно выполнена."
      security:
        - bearerAuth: []
  /auth/me:
    get:
      tags:
        - "auth"
      summary: "Текущий аутентифицированный пользователь"
      description: "Возвращает профиль текущего пользователя и его актуальные права (RBAC)"
      parameters: []
      responses:
        200:
          description: "models.CurrentUserResponse"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUserResponse"
              example:
                success: true
                data:
                  id: "usr-uuid-1"
                  email: "admin@course-platform.local"
                  full_name: "Администратор Системы"
                  role:
                    id: "role-uuid-admin"
                    name: "admin"
                    description: "Полный администратор платформы"
                    permissions:
                      - "users:view_all"
                      - "users:manage_roles"
                      - "roles:manage"
                      - "courses:create"
                      - "payments:view_all"
                  created_at: "2026-10-04T00:00:00.000Z"
        401:
          description: "models.ErrorResponse"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                error:
                  code: "VALIDATION_ERROR"
                  message: "Неверные параметры запроса."
      security:
        - bearerAuth: []
...